requirements.txt
Flask
requests
flask-session
werkzeug.security